Move files to follow new file layout conventions starting with Go1 and update Makefiles appropriately
New commands with Go1 include go build, go install, and go test

All new commands assume a GOPATH variable is set and that the current directory rests on an element of that path.  All *.go files must be under a src/ directory.

The generated library file is now stored to lib/go/pkg/*/thrift.a