Use local tests in Travis while adding build.
Travis CI seems to update its test image frequently, and it appears,
for instance, that the team is adding native Go support.  To that
end, I will relax the removal requirement of it, putting a preference
on the release that we locally install.

Eventually, this will be replaced by just purely a ``make test``
integration test target but not this second; and with this, we'll
simply have Travis CI invoke the makefile for testing versus
maintaining duplicate rules.