Fix path in the support library copying process.
The copying process for the thrift4go support library made an
erroneous assumption in that it assumed the GOPATH-oriented
directories should be created, not potentially ones that used the
``DESTDIR`` distribution packaging semantic.  This would have
hindered Linux users, for instance.  Plus, it fails with a test
that I am about to add.