Expand generation test to include default fields.
The generation test suite has been expanded to now include default
enum fields in the struct.  This will allow me to target the
presently-errant functionality with tests in a future commit.  To
achieve this effect, the test suite now includes a skeletal Go test
case that interacts with the generated Go code from a simple IDL.