Add test case validating optional field defaults.
Per https://github.com/pomack/thrift4go/issues/16, we'll need to
verify that the generated enum code really knows if a value or not
is set for such fields.  In reality, these tests pass but they really
hide the underlying error of the above issue.  They'll be changed to
fail on incorrect behavior, and the underlying generated code will be
fixed to make 'em pass in the next commit.