Validate is-set of enum field elements.
If an enum is unset and requested in string form, it emits `<UNSET>`
to provide a cue to the user.

The base value for being unset is outside of Thrift's enum signed
int32 value range.  The behavior around those enums that have default
values may need further verification.  This allows an enum field for
a struct that possesses unnumbered element values to be set to the
first value such that thrift4go now acknowledges that the the field
is actually set and will be included in the wire format.