Instruct the Go generator to run ``gofmt -w``.
The Go code generator from the Thrift IDL will now run ``gofmt -w``
on the emitted output; however, the Thrift generator will not
necessarily care if the formatting process succeeds.  This will
lower the barrier for entry for folks wanting to adopt Thrift on a
Go-based platform due to following preferred standards.