Add end-to-end interoperability test suite.
We now have a canonical interopability test suite that exercises the
behavior of the generated Go code against Java, which I perceive to
be the better-maintained Thrift languages available.  This will
enable us to verify wire protocol expectations between the two.

Already this commit uncovers a problem in TSimpleJSON, but it is
hard to tease apart whether this is necessarily new in thrift4go or
not.  It is presently commented out and will be investigated very
soon.

Right now Maven is used to build and host the Java sub-component.
Travis has been updated to install requisite dependencies, as I
expect them to be for this new stuff.

The interoperability tests could be made more robust in checking
a Java client to a Go server reliability, but this will come in the
near future.

As a side note, various ``Makefile`` entities across the repository
have been updated such that the test suites around them recognize
that they need to be re-run if the ``Makefile`` timestamp is newer
than the dependent's rule.  More re-workings can be done.