Add Cassandra Thrift IDLs for further validation.
Problems in the generated Go code for Cassandra from its IDL prompted
me to start making repairs to thrift4go.  That said, I want to add
some additional validations to validate that generation of the Go
code for these is actually buildable.  From this, a couple of
anomalies appeared in terms of how safe field names were
emitted-e.g., ``Type`` to a Go-safe name.