Travis config now works cross-repository
The .travis.yml config file had pomack/thrift4go statically in it, which made testing a commit from an other repository (fork) with travis unpossible.
I have added some lines that will figure out the name of the repository and will set that as an environmen var, which is now used troughout the config file (instead of hardcoded "pomack/thrift4go").